backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "passlib>=1.7.4",
    "pydantic-settings>=2.12.0",
    "redis>=7.1.0",
    "websockets>=15.0.1",
]
